requirements.txt
# Flask web framework
flask>=3.0.0
flask-cors>=4.0.0

# AI/OCR - Gemini (FREE tier!)
google-generativeai>=0.8.0
pillow>=10.0.0

# Google APIs (Calendar, Gmail)
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.111.0

# Cloud storage (Cloudflare R2)
boto3>=1.34.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.2.0
